Test structure (#22)
* Add test structure

* Run prettier

* Remove sesion state tests

* Run prettier
