+++ src/tests/utils/articleGenerator.ts
@@ -1,5 +1,4 @@
 import {RegularArticle} from '@app/article'
-import {User} from '@app/user'
 import {dummyAuthor1, dummyAuthor2, dummyAuthor3} from './dummies'
 
 const abstracts = [
